+++ HISTORY.rst
@@ -4,7 +4,8 @@ History
 
 2.0.1 (Unreleased)
 ------------------
-    * When transition happens we also notify also model updated event (if available) (rudaporto).
+    
+    * When transition happens we also notify model updated event (if available) (rudaporto).
 
 2.0.0 (2017-04-21)
 ------------------